Create TestTargetType protocol
