+++ Sources/Moya/MoyaProvider.swift
@@ -43,12 +43,8 @@ open class MoyaProvider<Target: TargetType>: MoyaProviderType {
     /// Closure that resolves an `Endpoint` into a `RequestResult`.
     public typealias RequestClosure = (Endpoint<Target>, @escaping RequestResultClosure) -> Void
 
-    /// Closure that decides if/how a request should be stubbed.
-    public typealias StubClosure = (Target) -> Moya.StubBehavior
-
     open let endpointClosure: EndpointClosure
     open let requestClosure: RequestClosure
-    open let stubClosure: StubClosure
     open let manager: Manager
 
     /// A list of plugins
@@ -65,7 +61,6 @@ open class MoyaProvider<Target: TargetType>: MoyaProviderType {
     /// Initializes a provider.
     public init(endpointClosure: @escaping EndpointClosure = MoyaProvider.defaultEndpointMapping,
                 requestClosure: @escaping RequestClosure = MoyaProvider.defaultRequestMapping,
-                stubClosure: @escaping StubClosure = MoyaProvider.neverStub,
                 callbackQueue: DispatchQueue? = nil,
                 manager: Manager = MoyaProvider<Target>.defaultAlamofireManager(),
                 plugins: [PluginType] = [],
@@ -73,7 +68,6 @@ open class MoyaProvider<Target: TargetType>: MoyaProviderType {
 
         self.endpointClosure = endpointClosure
         self.requestClosure = requestClosure
-        self.stubClosure = stubClosure
         self.manager = manager
         self.plugins = plugins
         self.trackInflights = trackInflights
@@ -94,71 +88,6 @@ open class MoyaProvider<Target: TargetType>: MoyaProviderType {
 
         let callbackQueue = callbackQueue ?? self.callbackQueue
         return requestNormal(target, callbackQueue: callbackQueue, progress: progress, completion: completion)
-    }
-
-    // swiftlint:disable function_parameter_count
-    /// When overriding this method, take care to `notifyPluginsOfImpendingStub` and to perform the stub using the `createStubFunction` method.
-    /// Note: this was previously in an extension, however it must be in the original class declaration to allow subclasses to override.
-    @discardableResult
-    open func stubRequest(_ target: Target, request: URLRequest, callbackQueue: DispatchQueue?, completion: @escaping Moya.Completion, endpoint: Endpoint<Target>, stubBehavior: Moya.StubBehavior) -> CancellableToken {
-        let callbackQueue = callbackQueue ?? self.callbackQueue
-        let cancellableToken = CancellableToken { }
-        notifyPluginsOfImpendingStub(for: request, target: target)
-        let plugins = self.plugins
-        let stub: () -> Void = createStubFunction(cancellableToken, forTarget: target, withCompletion: completion, endpoint: endpoint, plugins: plugins, request: request)
-        switch stubBehavior {
-        case .immediate:
-            switch callbackQueue {
-            case .none:
-                stub()
-            case .some(let callbackQueue):
-                callbackQueue.async(execute: stub)
-            }
-        case .delayed(let delay):
-            let killTimeOffset = Int64(CDouble(delay) * CDouble(NSEC_PER_SEC))
-            let killTime = DispatchTime.now() + Double(killTimeOffset) / Double(NSEC_PER_SEC)
-            (callbackQueue ?? DispatchQueue.main).asyncAfter(deadline: killTime) {
-                stub()
-            }
-        case .never:
-            fatalError("Method called to stub request when stubbing is disabled.")
-        }
-
-        return cancellableToken
-    }
-    // swiftlint:enable function_parameter_count
-}
-
-/// Mark: Stubbing
-
-/// Controls how stub responses are returned.
-public enum StubBehavior {
-
-    /// Do not stub.
-    case never
-
-    /// Return a response immediately.
-    case immediate
-
-    /// Return a response after a delay.
-    case delayed(seconds: TimeInterval)
-}
-
-public extension MoyaProvider {
-
-    // Swift won't let us put the StubBehavior enum inside the provider class, so we'll
-    // at least add some class functions to allow easy access to common stubbing closures.
-
-    public final class func neverStub(_: Target) -> Moya.StubBehavior {
-        return .never
-    }
-
-    public final class func immediatelyStub(_: Target) -> Moya.StubBehavior {
-        return .immediate
-    }
-
-    public final class func delayedStub(_ seconds: TimeInterval) -> (Target) -> Moya.StubBehavior {
-        return { _ in return .delayed(seconds: seconds) }
     }
 }
