+++ Sources/Moya/MultiTarget.swift
@@ -21,9 +21,10 @@ public enum MultiTarget: TargetType {
         return target.method
     }
 
-    public var sampleData: Data {
-        return target.sampleData
-    }
+// TODO: - Not sure how this would affect sample data
+//    public var sampleData: Data {
+//        return target.sampleData
+//    }
 
     public var task: Task {
         return target.task
