+++ Sources/Moya/TargetType.swift
@@ -12,9 +12,6 @@ public protocol TargetType {
     /// The HTTP method used in the request.
     var method: Moya.Method { get }
 
-    /// Provides stub data for use in testing.
-    var sampleData: Data { get }
-
     /// The type of HTTP task to be performed.
     var task: Task { get }
